i3t-demo (old contiki-ng repo for run rpl native)

+++ os/lib/cbor/cbor.h
@@ -49,11 +49,12 @@ int cbor_put_text(uint8_t **buffer, char *text, uint8_t text_len);
 
 int cbor_put_array(uint8_t **buffer, uint8_t elements);
 
-int cbor_put_bytes(uint8_t **buffer, uint8_t *bytes, uint8_t bytes_len);
+//int cbor_put_bytes(uint8_t **buffer, uint8_t *bytes, uint8_t bytes_len);
+int cbor_put_bytes(uint8_t **buffer, uint8_t *bytes, size_t bytes_len);
 
 int cbor_put_map(uint8_t **buffer, uint8_t elements);
 
-int cbor_put_unsigned(uint8_t **buffer, uint8_t value);
+int cbor_put_unsigned(uint8_t **buffer, size_t value);
 
 int cbor_put_negative(uint8_t **buffer, int64_t value);
